+++ backend/pipelines/lush_korea/__init__.py
@@ -1,0 +1,3 @@
+"""Lush Korea scraping modules."""
+
+# End of file.
