+++ backend/pipelines/lush_uk/__init__.py
@@ -1,0 +1,3 @@
+"""Lush UK scraping modules."""
+
+# End of file.
