+++ backend/tests/test_lush_uk_scraper.py
@@ -1,5 +1,11 @@
+"""Tests for the Lush UK scraper facade and modules."""
+
+from typing import Any
+
 import pytest
 
+from pipelines.lush_uk.detail import extract_product_detail as extract_uk_detail
+from pipelines.lush_uk.search import extract_search_products as extract_uk_search_products
 from pipelines.lush_uk_scraper import (
     DEFAULT_URL,
     extract_fragrance_products,
@@ -14,7 +20,7 @@ from pipelines.lush_uk_scraper import (
 )
 
 
-def test_extract_fragrance_products_from_lush_uk_category_html():
+def test_extract_fragrance_products_from_lush_uk_category_html() -> None:
     html = """
     <main>
       <article>
@@ -59,7 +65,7 @@ def test_extract_fragrance_products_from_lush_uk_category_html():
     ]
 
 
-def test_scrape_perfume_names_from_html_adds_uk_country():
+def test_scrape_perfume_names_from_html_adds_uk_country() -> None:
     rows = scrape_perfume_names_from_html(
         '<a href="/uk/en/p/karma-perfume"><h2>Karma</h2><p>Perfume</p></a>'
     )
@@ -74,7 +80,7 @@ def test_scrape_perfume_names_from_html_adds_uk_country():
     ]
 
 
-def test_extract_product_price_reads_visible_sterling_price():
+def test_extract_product_price_reads_visible_sterling_price() -> None:
     html = """
     <span class="sr-only">£30.00</span>
     <button>Add to bag - £30.00</button>
@@ -83,7 +89,7 @@ def test_extract_product_price_reads_visible_sterling_price():
     assert extract_product_price(html) == "£30.00"
 
 
-def test_extract_product_price_reads_next_data_variant_price():
+def test_extract_product_price_reads_next_data_variant_price() -> None:
     html = """
     <script id="__NEXT_DATA__" type="application/json">
     {
@@ -124,7 +130,9 @@ def test_extract_product_price_reads_next_data_variant_price():
     assert extract_product_price(html) == "£30.00"
 
 
-def test_scrape_perfume_names_from_non_default_url_adds_regular_price_from_product_page(monkeypatch):
+def test_scrape_perfume_names_from_non_default_url_adds_regular_price_from_product_page(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
     monkeypatch.setattr(
         "pipelines.lush_uk_scraper.fetch_category",
         lambda url: '<a href="/uk/en/p/dirty-perfume"><h2>Dirty</h2><p>Perfume</p></a>',
@@ -151,7 +159,7 @@ def test_scrape_perfume_names_from_non_default_url_adds_regular_price_from_produ
     ]
 
 
-def test_extract_search_products_reads_paginated_search_payload():
+def test_extract_search_products_reads_paginated_search_payload() -> None:
     payload = {
         "data": {
             "searchQuery": {
@@ -182,22 +190,53 @@ def test_extract_search_products_reads_paginated_search_payload():
     ]
 
 
-def test_fetch_collection_products_follows_search_pagination():
+def test_uk_search_module_exposes_search_product_extraction() -> None:
+    payload = {
+        "data": {
+            "searchQuery": {
+                "items": [
+                    {
+                        "content": {
+                            "name": "Dirty",
+                            "slug": "dirty-perfume",
+                            "minPrice": 30,
+                            "currency": "GBP",
+                            "attributes": {"type": "Perfume"},
+                        }
+                    }
+                ]
+            }
+        }
+    }
+
+    assert extract_uk_search_products(payload)[0]["product_url"] == (
+        "https://www.lush.com/uk/en/p/dirty-perfume"
+    )
+
+
+def test_fetch_collection_products_follows_search_pagination() -> None:
     class FakeResponse:
-        def __init__(self, payload):
+        def __init__(self, payload: dict[str, Any]) -> None:
             self._payload = payload
 
-        def raise_for_status(self):
+        def raise_for_status(self) -> None:
             return None
 
-        def json(self):
+        def json(self) -> dict[str, Any]:
             return self._payload
 
     class FakeSession:
-        def __init__(self):
+        def __init__(self) -> None:
             self.pages: list[int] = []
 
-        def post(self, _url, *, json, headers, timeout):
+        def post(
+            self,
+            _url: str,
+            *,
+            json: dict[str, Any],
+            headers: dict[str, str],
+            timeout: int,
+        ) -> FakeResponse:
             page = json["variables"]["page"]
             self.pages.append(page)
             return FakeResponse(
@@ -229,7 +268,7 @@ def test_fetch_collection_products_follows_search_pagination():
     assert [row["english_name"] for row in rows] == ["Product 1", "Product 2"]
 
 
-def test_extract_product_detail_reads_next_data_ingredients():
+def test_extract_product_detail_reads_next_data_ingredients() -> None:
     html = """
     <script id="__NEXT_DATA__" type="application/json">
     {
@@ -268,7 +307,17 @@ def test_extract_product_detail_reads_next_data_ingredients():
     }
 
 
-def test_scrape_perfume_names_adds_product_detail_fields(monkeypatch):
+def test_uk_detail_module_exposes_product_detail_extraction() -> None:
+    html = """
+    <script id="__NEXT_DATA__" type="application/json">
+    {"props":{"pageProps":{"__APOLLO_STATE__":{"Product:1":{"__typename":"Product","attributes":[{"attribute":{"slug":"key_ingredients"},"values":[{"name":"Mint"}]}]}}}}}
+    </script>
+    """
+
+    assert extract_uk_detail(html)["key_ingredients"] == ["Mint"]
+
+
+def test_scrape_perfume_names_adds_product_detail_fields(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.setattr(
         "pipelines.lush_uk_scraper.fetch_collection_products",
         lambda _collection_name: [
@@ -302,7 +351,7 @@ def test_scrape_perfume_names_adds_product_detail_fields(monkeypatch):
     ]
 
 
-def test_extract_referenced_products_from_next_data():
+def test_extract_referenced_products_from_next_data() -> None:
     html = """
     <script id="__NEXT_DATA__" type="application/json">
     {
@@ -344,7 +393,7 @@ def test_extract_referenced_products_from_next_data():
     ]
 
 
-def test_scrape_perfume_names_from_html_rejects_cloudflare_challenge():
+def test_scrape_perfume_names_from_html_rejects_cloudflare_challenge() -> None:
     with pytest.raises(RuntimeError, match="Cloudflare challenge"):
         scrape_perfume_names_from_html("Just a moment... Enable JavaScript and cookies to continue")
 
@@ -352,11 +401,11 @@ def test_scrape_perfume_names_from_html_rejects_cloudflare_challenge():
         scrape_perfume_names_from_html("<title>잠시만 기다리십시오…</title>challenges.cloudflare.com")
 
 
-def test_lush_uk_default_url_is_fragrance_category():
+def test_lush_uk_default_url_is_fragrance_category() -> None:
     assert DEFAULT_URL == "https://www.lush.com/uk/en/c/fragrances"
 
 
-def test_fetch_category_uses_scrapling_by_default(monkeypatch):
+def test_fetch_category_uses_scrapling_by_default(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.delenv("LUSH_UK_USE_SELENIUM", raising=False)
     monkeypatch.delenv("LUSH_UK_FETCHER", raising=False)
     monkeypatch.setattr(
@@ -367,12 +416,12 @@ def test_fetch_category_uses_scrapling_by_default(monkeypatch):
     assert fetch_category() == "<html>https://www.lush.com/uk/en/c/fragrances</html>"
 
 
-def test_fetch_category_can_use_requests(monkeypatch):
+def test_fetch_category_can_use_requests(monkeypatch: pytest.MonkeyPatch) -> None:
     class FakeResponse:
         encoding = ""
         text = "<html>requests</html>"
 
-        def raise_for_status(self):
+        def raise_for_status(self) -> None:
             return None
 
     monkeypatch.setenv("LUSH_UK_FETCHER", "requests")
@@ -384,10 +433,12 @@ def test_fetch_category_can_use_requests(monkeypatch):
     assert fetch_category() == "<html>requests</html>"
 
 
-def test_fetch_category_can_use_selenium_rendered_html(monkeypatch):
+def test_fetch_category_can_use_selenium_rendered_html(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.setattr(
         "pipelines.lush_uk_scraper.fetch_rendered_html",
         lambda url: f"<html>{url}</html>",
     )
 
     assert fetch_category(use_selenium=True) == "<html>https://www.lush.com/uk/en/c/fragrances</html>"
+
+# End of file.
